+++ examples/zstd_dict.c
@@ -19,7 +19,6 @@
 #include <stdio.h>
 #include <assert.h>
 #include <time.h>
-#include <blosc2.h>
 #include "blosc2.h"
 
 #define KB  1024.
@@ -34,7 +33,7 @@
 int main(void) {
   static int64_t data[CHUNKSIZE];
   static int64_t data_dest[CHUNKSIZE];
-  const size_t isize = CHUNKSIZE * sizeof(int64_t);
+  const int32_t isize = CHUNKSIZE * sizeof(int64_t);
   int dsize = 0;
   int64_t nbytes, cbytes;
   blosc2_schunk* schunk;
